refactor(settings): move currency setting to inventory manager page
