+++ docs/changelogs/5.3.10.mdx
@@ -1,13 +1,16 @@
 ---
 title: 5.3.10
 date: 2026-06-26
-tags: ['New', 'Fixed']
+tags: ['New', 'Improved', 'Fixed']
 ---
 
 ### New 🎉
 - **New PRO Gamer Tier Feature**: Added a **Run Multiple Games** setting to **Settings > Achievement Unlocker**
   - When enabled, SGI unlocks achievements for up to 32 games at the same time instead of one at a time, each with its own independent unlock schedule
 - The **Achievement Unlocker** now shows a queue of the next 5 upcoming achievements, each with its own live countdown until unlock
+
+### Improved 🚀
+- Moved the **Currency** setting from **Settings > General** to **Settings > Inventory Manager** as this setting is only used by that feature
 
 ### Fixed 🐛
 - Fixed an issue in **Inventory Manager** where Steam credentials were not being automatically revalidated for **PRO** users
