+++ src/features/settings/components/inventory-manager/CurrencySwitch.tsx
